Add files via upload
The NIR autoEncoder encodes the 700 bands of the NIR spectroscopy into a smaller dimensional space such as two bands and then trains to minimize recreation error. This plot is the 2d representation of the data. The groups are expiremental groups that the NIR data came from. The different groups analyzed wheat from different locations with a different technicians and/or NIR machines. 

The data clearly split into two clusters. Group A was 1 cluster while group B  & C were another cluster.  My office Mate used a CNN with limited success to predict the Falling number of the wheat flour. I think that the groups having different locations and technicians/machines is influncing the features of the NIR series more then the falling number confounding the prediction.